Refactor: Module 수정 #64 UserRepositoryModule에 있는 SharedPreferences 모듈을 최상위에 위치한 AppModule에 위치로 변경
